+++ base-compat/src/Data/List/Compat.hs
@@ -5,6 +5,7 @@ module Data.List.Compat (
   module Base
 
 #if !(MIN_VERSION_base(4,21,0))
+, compareLength
 , inits1
 , tails1
 #endif
@@ -29,21 +30,14 @@ module Data.List.Compat (
 
 import Data.List as Base
 
-#if !(MIN_VERSION_base(4,11,0))
-import GHC.Exts (build)
-#endif
-
-#if !(MIN_VERSION_base(4,19,0))
-import Prelude.Compat hiding (foldr, null)
-#endif
-
 #if MIN_VERSION_base(4,18,0) && !(MIN_VERSION_base(4,20,0))
 import GHC.List (List)
 #endif
 
 #if !(MIN_VERSION_base(4,21,0))
 import Data.List.NonEmpty (NonEmpty(..))
-import GHC.List (build)
+import GHC.Exts (build)
+import Prelude.Compat hiding (foldr, null)
 #endif
 
 #if !(MIN_VERSION_base(4,11,0))
@@ -145,6 +139,40 @@ unsnoc = foldr (\x -> Just . maybe ([], x) (\(~(a, b)) -> (x : a, b))) Nothing
 #endif
 
 #if !(MIN_VERSION_base(4,21,0))
+-- | Use 'compareLength' @xs@ @n@ as a safer and faster alternative
+-- to 'compare' ('length' @xs@) @n@. Similarly, it's better
+-- to write @compareLength xs 10 == LT@ instead of @length xs < 10@.
+--
+-- While 'length' would force and traverse
+-- the entire spine of @xs@ (which could even diverge if @xs@ is infinite),
+-- 'compareLength' traverses at most @n@ elements to determine its result.
+--
+-- >>> compareLength [] 0
+-- EQ
+-- >>> compareLength [] 1
+-- LT
+-- >>> compareLength ['a'] 1
+-- EQ
+-- >>> compareLength ['a', 'b'] 1
+-- GT
+-- >>> compareLength [0..] 100
+-- GT
+-- >>> compareLength undefined (-1)
+-- GT
+-- >>> compareLength ('a' : undefined) 0
+-- GT
+--
+-- @since 4.21.0.0
+--
+compareLength :: [a] -> Int -> Ordering
+compareLength xs n
+  | n < 0 = GT
+  | otherwise = foldr
+    (\_ f m -> if m > 0 then f (m - 1) else GT)
+    (\m -> if m > 0 then LT else EQ)
+    xs
+    n
+
 inits1, tails1 :: [a] -> [NonEmpty a]
 
 -- | The 'inits1' function returns all non-empty initial segments of the